Moved Clours.ts to /src. Not a component
